% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPercPyramid.R
\name{plotPercPyramid}
\alias{plotPercPyramid}
\title{Population percentage pyramid graph}
\usage{
plotPercPyramid(
  df,
  age,
  group,
  perpop,
  labx = perpop,
  laby = age,
  twocolors = c("#1b9e77", "#7570b3"),
  rotation = 0,
  n.breaks = 20,
  value.labels = TRUE,
  position.value.labels = "in",
  size.value.labels = 3
)
}
\arguments{
\item{df}{Name of dataframe}

\item{age}{Age or age group. Write the parameter in quotation marks.}

\item{group}{Categorical grouping variable. Write the parameter in quotation marks.}

\item{perpop}{Percentage of population (in numerical value). Write the parameter in quotation marks.}

\item{labx}{X-axis label}

\item{laby}{Y-axis label}

\item{twocolors}{Two colors for the pyramid}

\item{rotation}{X-axis label rotation}

\item{n.breaks}{Number of breaks}

\item{value.labels}{Show values in the bars. Use TRUE to include the labels in the bars. Use FALSE to not include them.}

\item{position.value.labels}{Position of the values on the bars. Use "in" to display the labels inside the bars. Use "out" to display them outside the bars.}

\item{size.value.labels}{Font size of the values in the bars}
}
\value{
A graph of the pyramid of population percentage
}
\description{
Create a population percentage pyramid graph
}
\examples{
library(readxl)
url <- "https://zenodo.org/record/8306939/files/popPER.xlsx?download=1"
destfile <- "popPER.xlsx"
curl::curl_download(url, destfile)
data <- read_excel(destfile)
data <- dplyr::filter(data, Year==2023)
library(dplyr)
data <- data |>
group_by(gAge, Sex) |>
summarise(Population = sum(Population)) |>
group_by(Sex) |>
mutate(percPopulation = round(Population/sum(Population)*100, 4)) |>
select(gAge, Sex, percPopulation) |>
arrange(Sex, gAge)
plotPercPyramid(df=data, age="gAge", group="Sex", perpop="percPopulation", value.labels=T)

}
