% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPyramid.R
\name{plotPyramid}
\alias{plotPyramid}
\title{Population pyramid graph}
\usage{
plotPyramid(
  df,
  age,
  group,
  pop,
  labx = pop,
  laby = age,
  twocolors = c("#1b9e77", "#7570b3"),
  rotation = 90,
  n.breaks = 20,
  value.labels = TRUE,
  position.value.labels = "in",
  size.value.labels = 3
)
}
\arguments{
\item{df}{Name of dataframe}

\item{age}{Age or age group. Write the parameter in quotation marks.}

\item{group}{Categorical grouping variable. Write the parameter in quotation marks.}

\item{pop}{Population (in numerical value). Write the parameter in quotation marks.}

\item{labx}{X-axis label}

\item{laby}{Y-axis label}

\item{twocolors}{Two colors for the pyramid}

\item{rotation}{X-axis label rotation}

\item{n.breaks}{Number of breaks}

\item{value.labels}{Show values in the bars. Use TRUE to include the labels in the bars. Use FALSE to not include them.}

\item{position.value.labels}{Position of the values on the bars. Use "in" to display the labels inside the bars. Use "out" to display them outside the bars.}

\item{size.value.labels}{Font size of the values in the bars}
}
\value{
A population pyramid graph
}
\description{
Create a population pyramid graph
}
\examples{
library(readxl)
url <- "https://zenodo.org/record/8306939/files/popPER.xlsx?download=1"
destfile <- "popPER.xlsx"
curl::curl_download(url, destfile)
data <- read_excel(destfile)
data <- dplyr::filter(data, Year==2023)
plotPyramid(df=data, age="gAge", group="Sex", pop="Population", value.labels=T)

}
