% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpyramid.R
\name{ggpyramid}
\alias{ggpyramid}
\title{Function for plotting population pyramids with ggplot2}
\usage{
ggpyramid(
  data = data,
  age = age,
  pop = pop,
  gender = gender,
  men = "Men",
  women = "Women",
  fill = gender
)
}
\arguments{
\item{data}{Data frame with population data by age and gender}

\item{age}{Column with the age groups}

\item{pop}{Column with population counts}

\item{gender}{Column with gender data}

\item{men}{How males are define in "gender" (default = "Men")}

\item{women}{How females are defined in "gender" (default = "Women")}

\item{fill}{Fill bars based on another variable (default = gender)}
}
\value{
population pyramid
}
\description{
Function for plotting population pyramids with ggplot2
}
\examples{
library(tidyverse)
library(ggpyramid)
library(danstat)

# Load data from Denmark using (danstat)
var_input <- list(list(code = "OMRÅDE", values = "000"),
                  list(code = "KØN", values = c(1, 2)),
                  list(code = "ALDER", values = seq(0, 125, 1)),
                  list(code = "tid", values = "2020K4"))

get_data("FOLK1A", variables = var_input) \%>\%
  rename(country = OMRÅDE,
         gender = KØN,
         age = ALDER,
         year = TID,
         pop = INDHOLD) \%>\%
  mutate(age = parse_number(age),
         age = as.integer(age)) -> df

df \%>\%
  ggpyramid(pop = pop, fill = gender) +
  scale_fill_manual(name = "Gender", values = c("#0072B2", "#F0E442")) +
  labs(title = "Population pyramith of Denmark",
       caption = "Source: Statistics Denmark")

}
