% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpyramid.R
\name{ggpyramid}
\alias{ggpyramid}
\title{Age pyramid using ggplot2}
\usage{
ggpyramid(data, age, sex, males = stringr::regex("1|^m.*", ignore_case =
  TRUE), females = stringr::regex("2|^f.*", ignore_case = TRUE),
  x_breaks, make_age_cat = TRUE)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{age}{...}

\item{sex}{...}

\item{males}{...}

\item{females}{...}

\item{x_breaks}{...}

\item{make_age_cat}{...}
}
\description{
This is where you put your description. Bla bla bla. Lines should not be more
than 80 characters.
}
\examples{
age <- rgeom(1000, .1)
sex <- sample(c("male", "female"), 1000, replace = TRUE)
df <- data.frame(age, sex)
ggpyramid(df, age = age, sex = sex, x_breaks = 50)
}
\author{
Jonathan Polonsky \email{polonskyj@who.int}.
}
